layer.toDataset(): Layer transfrom to Dataset
